refactor(comment): emotion selector re-org
