+++ src/components/EmotionSelector/SelectedEmotions/EmotionUnit.tsx
@@ -1,0 +1,56 @@
+/*
+ * EmojiSelector
+ */
+
+import { FC, memo } from 'react'
+import { buildLog } from '@/utils/logger'
+
+import type { TEmotion, TSimpleUser, TEmotionType } from '@/spec'
+
+import { titleCase } from '@/utils/helper'
+import Tooltip from '@/components/Tooltip'
+import AnimatedCount from '@/components/AnimatedCount'
+
+import EmotionIcon from './EmotionIcon'
+import UsersPanel from './UsersPanel'
+import { getEmotionName } from './helper'
+
+import { Wrapper, Count } from '../styles/users_panel'
+
+/* eslint-disable-next-line */
+const log = buildLog('c:EmotionUnit:index')
+
+type TProps = {
+  item: TEmotion
+  onAction?: (name: TEmotionType, hasEmotioned: boolean) => void
+}
+
+const EmotionUnit: FC<TProps> = ({ item, onAction }) => {
+  const name = getEmotionName(item)
+  const count = item[`${name}Count`] as number
+  const users = item[`latest${titleCase(name)}Users`] as TSimpleUser[]
+  const hasEmotioned = item[`viewerHas${titleCase(name)}ed`] as boolean
+
+  return (
+    <Tooltip
+      content={
+        <UsersPanel
+          name={name}
+          count={count}
+          users={users}
+          hasEmotioned={hasEmotioned}
+        />
+      }
+      noPadding
+    >
+      <Wrapper onClick={() => onAction(name as TEmotionType, hasEmotioned)}>
+        <EmotionIcon name={name} />
+        <Count>
+          <AnimatedCount count={count} size="tiny" active={false} />
+        </Count>
+      </Wrapper>
+    </Tooltip>
+  )
+}
+
+export default memo(EmotionUnit)
